refactor: obj, gltf 로드 분리
